requirements.txt
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-storage==3.4.0
google-cloud-vision==3.10.2
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
# grpcio==1.75.0
# grpcio-status==1.75.0
httplib2==0.31.0
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
requests==2.32.5
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
